Moved build context to efficiency module
