+++ docker/errors.py
@@ -86,3 +86,8 @@ class TLSParameterError(DockerException):
 
 class NullResource(DockerException, ValueError):
     pass
+
+
+class ContextError(DockerException):
+    def __init__(self, msg):
+        self.message = msg
