+++ docker/utils/__init__.py
@@ -8,8 +8,3 @@ from .utils import (
 
 from .types import Ulimit, LogConfig # flake8: noqa
 from .decorators import check_resource, minimum_version #flake8: noqa
-from .context import (
-    create_context_from_path,
-    is_remote,
-    ContextError
-) # flake8: noqa
